requirements.txt
aiohttp
requests
websockets
python-dotenv
pytz
ccxt
ta
pandas
numpy
scipy
python-telegram-bot==13.15
